requirements.txt
Pillow
qrcode
